% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_units.R
\name{convert_units}
\alias{convert_units}
\title{Convert units}
\usage{
convert_units(value, what, from, to)
}
\arguments{
\item{value}{numeric. Value to be converted}

\item{what}{character. What to convert}

\item{from}{character. Unit of the value}

\item{to}{character. Unit to convert to}
}
\value{
numeric. Converted value
}
\description{
Convert units of a value to another unit, conversions are based on the type of value
}
\details{
This function converts units of a value to another unit, conversions are based on the type of value:\tabular{ll}{
   Type \tab Units \cr
   temperature \tab celsius, fahrenheit, kelvin \cr
   pressure \tab atm, mmHg, kPa, bar, mbar, psi, mH2O \cr
   mass \tab kg, g, mg, lb, oz \cr
   velocity \tab km/h, m/s, mph \cr
   radiation \tab W/m2, cal/cm2, MJ/m2 \cr
   daily radiation \tab MJ/m2/day, W/m2/day, cal/cm2/day \cr
   time \tab s, min, h, d \cr
   length \tab m, cm, mm, km, in, ft, yd, mi \cr
   area \tab m2, cm2, mm2, km2, in2, ft2, yd2, mi2 \cr
   volume \tab m3, cm3, mm3, l, in3, ft3, yd3, gal \cr
   angle \tab deg, rad, grad, turn \cr
   conductivity \tab dS/m, mS/m, uS/m, mS/cm, uS/cm, tdi mg/L, Meq/l \cr
   density \tab kg/m3, g/cm3, g/m3, mg/m3, lb/ft3, lb/in3, lb/gal, lb/yd3, ton/yd3 \cr
}
}
\examples{
convert_units(1, "pressure", "atm", "kPa")
convert_units(101.325, "pressure", "kPa", "atm")
convert_units(1, "pressure", "atm", "bar")
convert_units(1.01325, "pressure", "bar", "atm")
convert_units(1, "pressure", "atm", "psi")
convert_units(14.6959, "pressure", "psi", "atm")
convert_units(1, "pressure", "atm", "mH2O")

}
