% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dailychillinghours.R
\name{dailychillinghours}
\alias{dailychillinghours}
\title{Estimate chilling hours from daily data}
\usage{
dailychillinghours(
  t_max,
  t_min,
  kf = 1,
  threshold = 7.2,
  t_maxu = "celsius",
  t_minu = "celsius",
  thresholdunit = "celsius",
  convert_to_celsius = TRUE
)
}
\arguments{
\item{t_max}{numeric. Maximum temperature.}

\item{t_min}{numeric. Minimum temperature.}

\item{kf}{numeric. Coefficient (from 1 to 1.5 recommended)}

\item{threshold}{numeric. Threshold temperature (default is 7.2)}

\item{t_maxu}{character. Unit of maximum temperature (celsius, fahrenheit, kelvin)}

\item{t_minu}{character. Unit of minimum temperature (celsius, fahrenheit, kelvin)}

\item{thresholdunit}{character. Unit of threshold temperature (celsius, fahrenheit, kelvin)}

\item{convert_to_celsius}{logical. Convert temperatures to celsius (default is TRUE)}
}
\value{
numeric. Daily chilling hours.
}
\description{
This functions estimates the daily chilling hours based on the maximum and minimum temperature data. The chilling hours are calculated based on a threshold temperature. The default threshold temperature is 7.2ºC.
}
\examples{
dailychillinghours(20, 6, kf = 1, threshold = 7.2)

dailychillinghours(20, 42.8, kf = 1, threshold = 45, t_minu = "fahrenheit",
                   thresholdunit = "fahrenheit", convert_to_celsius = TRUE)

}
