% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/et0h.R
\name{et0h}
\alias{et0h}
\title{Compute reference evapotranspiration using Hargreaves method}
\usage{
et0h(
  date,
  t_max,
  t_min,
  lat,
  t_max_u = "celsius",
  t_min_u = "celsius",
  lat_u = "rad"
)
}
\arguments{
\item{date}{character or Date object. Date of the observation.}

\item{t_max}{numeric. Maximum temperature in Celsius by default.}

\item{t_min}{numeric. Minimum temperature in Celsius by default.}

\item{lat}{numeric. Latitude of the location in decimal degrees by default.}

\item{t_max_u}{character. Unit of the maximum temperature. Default is "celsius".}

\item{t_min_u}{character. Unit of the minimum temperature. Default is "celsius".}

\item{lat_u}{character. Unit of the latitude. Default is "deg".}
}
\value{
A numeric value of the reference evapotranspiration in mm/day.
}
\description{
Compute daily reference evapotranspiration using Hargreaves method.
This function also allows different units for the input variables. For
unit conversion see \code{convert_units()} function.
}
\examples{
et0h(date='2024-03-09',t_max=27.69,t_min=7.91,lat=-30.631261)

}
