% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/et0pan.R
\name{et0pan}
\alias{et0pan}
\title{Compute daily reference evapotranspiration using Pav evaporation method.}
\usage{
et0pan(
  epan,
  kp = NULL,
  calculate_kp = FALSE,
  u2 = NULL,
  rh_mean = NULL,
  fet = NULL,
  class = c("class a", "colorado"),
  fetch = c("green", "dry")
)
}
\arguments{
\item{epan}{numeric. Pan evaporation in mm/day.}

\item{kp}{numeric or NULL. Coefficient of the pan. Default is NULL.}

\item{calculate_kp}{logical. If TRUE, kp will be calculated using the formula proposed by Allen et al. (1998). Default is FALSE.}

\item{u2}{numeric or NULL. Wind speed at 2 meters in m/s. Default is NULL.}

\item{rh_mean}{numeric or NULL. Mean relative humidity in \%. Default is NULL.}

\item{fet}{numeric or NULL. Fetch length in meters. Default is NULL.}

\item{class}{character. Class of the pan. Default is "class a".}

\item{fetch}{character. Fetch of the pan. Default is "green".}
}
\value{
A numeric value of the reference evapotranspiration in mm/day.
}
\description{
Compute daily reference evapotranspiration using Pav evaporation method from FAO-56
}
\details{
Measure water evaporated from a Pan to estimate reference evapotranspiration (ETo) using the FAO-56 Pan evaporation method. This function requires pan evaporation and pan coefficient (Kp). Kp can be estimated using mean wind speed at 2 meters, mean relative humidity, fetch length, class of the pan and fetch of the pan. There are 2 different classes of pan:
Class A and Class B. Fetch of the pan can be green or dry. The fetch is the ground cover around the pan.
}
\examples{
# FAO-56 data example:
# pan evaporation: 7.9 mm/day
# Fetch = 1000 m
# wind speed at 2 meters: 1.9 m/s
# mean relative humidity: 73\%
# Solution (in order): 6.6, 4.8, 7.7 ad 5.4 mm/day
et0pan(epan = 7.9,
       calculate_kp = TRUE,
       u2 = 1.9,
       fet = 1000,
       rh_mean = 73,
       class = 'class a',
       fetch = 'green')

et0pan(epan = 7.9,
       calculate_kp = TRUE,
       u2 = 1.9,
       fet = 1000,
       rh_mean = 73,
       class = 'class a',
       fetch = 'dry')

et0pan(epan = 7.9,
       calculate_kp = TRUE,
       u2 = 1.9,
       fet = 1000,
       rh_mean = 73,
       class = 'colorado',
       fetch = 'green')

et0pan(epan = 7.9,
       calculate_kp = TRUE,
       u2 = 1.9,
       fet = 1000,
       rh_mean = 73,
       class = 'colorado',
       fetch = 'dry')

}
