% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/et0pm.R
\name{et0pm}
\alias{et0pm}
\title{Compute reference evapotranspiration using Penman-Monteith method}
\usage{
et0pm(
  date,
  lat,
  elev,
  t_max,
  t_min,
  u_h,
  r_s,
  rh_max = NULL,
  rh_min = NULL,
  rh_mean = NULL,
  h = 2,
  lat_u = "deg",
  elev_u = "m",
  t_max_u = "celsius",
  t_min_u = "celsius",
  u_h_u = "m/s",
  r_s_u = "W/m2/day",
  rh_max_u = "\%",
  rh_min_u = "\%",
  rh_mean_u = "\%",
  h_u = "m"
)
}
\arguments{
\item{date}{character or Date object. Date of the observation.}

\item{lat}{numeric. Latitude of the location in decimal degrees by default.}

\item{elev}{numeric. Elevation of the location in meters by default.}

\item{t_max}{numeric. Maximum temperature in Celsius by default.}

\item{t_min}{numeric. Minimum temperature in Celsius by default.}

\item{u_h}{numeric. Wind speed at 2 meters in m/s by default.}

\item{r_s}{numeric. Solar radiation in W/m2/day by default.}

\item{rh_max}{numeric. Maximum relative humidity in \% by default.}

\item{rh_min}{numeric. Minimum relative humidity in \% by default.}

\item{rh_mean}{numeric. Mean relative humidity in \% by default.}

\item{h}{numeric. Height of the wind speed measurement in meters by default.}

\item{lat_u}{character. Unit of the latitude. Default is "deg".}

\item{elev_u}{character. Unit of the elevation. Default is "m".}

\item{t_max_u}{character. Unit of the maximum temperature. Default is "celsius".}

\item{t_min_u}{character. Unit of the minimum temperature. Default is "celsius".}

\item{u_h_u}{character. Unit of the wind speed. Default is "m/s".}

\item{r_s_u}{character. Unit of the solar radiation. Default is "W/m2/day".}

\item{rh_max_u}{character. Unit of the maximum relative humidity. Default is "\%".}

\item{rh_min_u}{character. Unit of the minimum relative humidity. Default is "\%".}

\item{rh_mean_u}{character. Unit of the mean relative humidity. Default is "\%".}

\item{h_u}{character. Unit of the height of the wind speed measurement. Default is "m".}
}
\value{
A numeric value of the reference evapotranspiration in mm/day.
}
\description{
Compute daily reference evapotranspiration using Penman-Monteith method.
This function also allows different units for the input variables. For
unit conversion see \code{convert_units()} function.
}
\details{
Calculate the reference evapotranspiration (ETo) using the FAO-56 Penman-Monteith method. This functions requires latitude, elevation, date, maximum temperature, minimum temperature, mean wind speed and mean solar radiation. There are 4 different methods to calculate actual vapor pressure, which are:
\itemize{
\item 1st method: using maximum and minimum Relative Humidity (\%).
\item 2nd method: using maximum Relative Humidity (\%).
\item 3rd method: using mean Relative Humidity (\%).
\item 4th method: using minimum temperature, asssuming when temperature is close to t_min, the air is (or close to) saturated.
}
}
\examples{
et0pm(date='2024-03-09',lat=-30.631261,elev=80,t_min=7.91,
      t_max=27.69,rh_min=37.71,rh_max=93.2,
      u_h=1.364,r_s=272.824)

et0pm(date='2024-03-09',lat=-30.631261,elev=80,t_min=281.06,
      t_max=300.84,rh_min=37.71,rh_max=93.2,
      u_h=1.364,r_s=272.824, t_min_u = "kelvin",
      t_max_u = "kelvin")

}
