% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraterrestrial_radiation.R
\name{extraterrestrial_radiation}
\alias{extraterrestrial_radiation}
\title{Calculate the extraterrestrial radiation.}
\usage{
extraterrestrial_radiation(j, lat)
}
\arguments{
\item{j}{numeric. Day of the year.}

\item{lat}{numeric. Latitude in radians.}
}
\value{
A numeric value of the clear sky radiation in MJ/m2/day.
}
\description{
This function computes the extraterrestrial radiation using the formula proposed by Allen et al. (1998).
}
\examples{
# date='2024-03-09'
# lat=-30.631261

lat = convert_units(-30.631261, "angle", "deg", "rad")
extraterrestrial_radiation(j=69, lat=lat)

}
