% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdd.R
\name{gdd}
\alias{gdd}
\title{Growing Degree Days (HDD) calculation}
\usage{
gdd(
  t_max = NULL,
  t_min = NULL,
  t_mean = NULL,
  base = 10,
  cutoff = 30,
  tunit = "celsius",
  lunit = "celsius"
)
}
\arguments{
\item{t_max}{numeric, maximum temperature}

\item{t_min}{numeric, minimum temperature}

\item{t_mean}{numeric, mean temperature}

\item{base}{numeric, base temperature}

\item{cutoff}{numeric, upper threshold temperature}

\item{tunit}{character, temperature unit}

\item{lunit}{character, length unit}
}
\value{
numeric, Growing Degree Days
}
\description{
This function calculates the Growing Degree Days (HDD) based on the mean temperature. The unit can be in Celsius, Fahrenheit, or Kelvin.
}
\details{
The Growing Degree Days (HDD) is a measure of heat accumulation used in agriculture to predict the growth and development of plants and insects. The HDD is calculated based on the mean temperature and the base temperature. The base temperature is the minimum temperature at which a plant or insect can grow. The default base temperature is 10ºC.
}
\examples{
gdd(t_max = 20, t_min = 10, tunit = 'celsius', lunit = 'celsius')
gdd(t_max=40, t_min=20)
gdd(t_max=40, t_min=20, cutoff=NULL)

}
