% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdd.R
\name{hdd}
\alias{hdd}
\title{Heating Degree Days (HDD) calculation}
\usage{
hdd(
  t_max = NULL,
  t_min = NULL,
  t_mean = NULL,
  base = 18.33,
  tunit = "celsius",
  lunit = "celsius"
)
}
\arguments{
\item{t_max}{numeric, maximum temperature}

\item{t_min}{numeric, minimum temperature}

\item{t_mean}{numeric, mean temperature}

\item{base}{numeric, base temperature}

\item{tunit}{character, temperature unit}

\item{lunit}{character, length unit}
}
\value{
numeric, Heating Degree Days
}
\description{
This function calculates the Heating Degree Days (HDD) based on the mean temperature. The unit can be in Celsius, Fahrenheit, or Kelvin.
}
\details{
The Heating Degree Days (HDD) is a measure of how much (in degrees), and for how long (in days), outside air temperature was lower than a specific base temperature. The base temperature is typically 18.33 degrees Celsius (65 degrees Fahrenheit).
The original formula is calculating t_mean with t_max and t_min, but t_mean option is added in the case you have only t_mean.
}
\examples{
hdd(t_max = 20, t_min = 10, tunit = 'celsius', lunit = 'celsius')
hdd(t_max = 30, t_min = 15, tunit = 'celsius', lunit = 'celsius')
hdd(t_max = 35, t_min = 18, tunit = 'celsius', lunit = 'celsius')

}
