% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hourly_to_daily.R
\name{hourly_to_daily}
\alias{hourly_to_daily}
\title{Convert hourly data to daily data}
\usage{
hourly_to_daily(
  df,
  datecol = "date",
  func = "mean",
  use_suffix = FALSE,
  no_suffix_func = "sum",
  na.rm = TRUE,
  func_suffix = c("min", "max", "mean", "sum", "sd", "var", "n")
)
}
\arguments{
\item{df}{data.frame, the data frame with hourly data}

\item{datecol}{character, the name of the column with the date}

\item{func}{character, the function to calculate the daily data, it can be 'min', 'max', 'mean' or any other function}

\item{use_suffix}{logical, if TRUE the function will calculate the daily data based on the suffix of each column}

\item{no_suffix_func}{character, the function to calculate the daily data when the column has no suffix}

\item{na.rm}{logical, if TRUE the function will remove NA values}

\item{func_suffix}{character, the suffix of each function to identify. By default these are: 'min', 'max', 'mean', 'sum', 'sd', 'var', 'n'}
}
\value{
data.frame, the data frame with daily data
}
\description{
This function converts hourly data to daily data. The function can calculate the daily data based on the mean, sum, min, max, standard deviation, variance, and number of observations based on the suffix of each column.
}
\examples{
data("piscoelqui")
df = piscoelqui$data
df = hourly_to_daily(df, date = "date", use_suffix = TRUE)

df = hourly_to_daily(df, date = "date", func = "mean", use_suffix = FALSE)

}
