% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chillinghours.R
\name{chillinghours}
\alias{chillinghours}
\title{Compute chilling hours}
\usage{
chillinghours(
  t_h,
  threshold = c(0, 7.2),
  tunit = "celsius",
  thresholdunit = "celsius",
  output = c("chillhours", "chillunits"),
  na.rm = TRUE
)
}
\arguments{
\item{t_h}{numeric. A vector of daily temperature data.}

\item{threshold}{numeric. A vector of two values representing the lower and upper threshold temperatures.}

\item{tunit}{character. The unit of the temperature data. Default is "celsius".}

\item{thresholdunit}{character. The unit of the threshold temperature. Default is "celsius".}

\item{output}{character. The type of output. Default is "chillhours".}

\item{na.rm}{logical. Should missing values be removed? Default is TRUE.}
}
\value{
A numeric value representing the chilling hours or chilling units.
}
\description{
Function to compute chilling hours based on the daily temperature data.
}
\examples{
chillinghours(c(5, 6, 7, 8, 9, 10))

}
