% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clearsky_radiation.R
\name{clearsky_radiation}
\alias{clearsky_radiation}
\title{Calculate the clear sky radiation.}
\usage{
clearsky_radiation(ra, elev)
}
\arguments{
\item{ra}{numeric. Extraterrestrial radiation in MJ/m2/day.}

\item{elev}{numeric. Elevation in meters.}
}
\value{
A numeric value of the clear sky radiation in MJ/m2/day.
}
\description{
This function computes the clear sky radiation using the formula proposed by Allen et al. (1998).
}
\examples{
clearsky_radiation(ra=35.16739, elev=80)

}
